test: refactor: split ha_cluster_info tests
